test: add custom commands for sign up
