+++ cypress/integration/user.ts
@@ -7,12 +7,7 @@ describe('User', () => {
     const user = UserFactory.build()
 
     cy.visit('/sign-up')
-
-    cy.findByPlaceholderText(/Username/i).type(user.username)
-    cy.findByPlaceholderText(/email/i).type(user.email)
-    cy.findByPlaceholderText(/^password/i).type(user.password)
-    cy.findByPlaceholderText(/confirm password/i).type(user.password)
-    cy.findByRole('button', { name: /Sign up now/i }).click()
+    cy.signUp(user)
 
     cy.wait(300)
