+++ cypress/support/factories.ts
@@ -1,11 +1,7 @@
 import { Factory } from 'rosie'
 import faker from '@faker-js/faker'
 
-type User = {
-  username: string
-  email: string
-  password: string
-}
+import type { User } from './types'
 
 export const UserFactory = new Factory<User>()
   .attr('username', () => faker.name.firstName().toLowerCase())
