+++ cypress/support/types.ts
@@ -1,0 +1,15 @@
+export type Field = {
+  name: string | number
+  label: string
+}
+
+export type ShowCase = {
+  name: string
+  highlight?: boolean
+}
+
+export type User = {
+  username: string
+  email: string
+  password: string
+}
